# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "API Documentation System Management",
        "termsOfService": "https://github.com/ducvubo",
        "contact": {
            "name": "Vũ Đức Bo",
            "url": "https://github.com/ducvubo",
            "email": "vminhduc8@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "system.management.pg.taphoaictu.id.vn",
    "basePath": "/api/v1",
    "paths": {
...
        "/internal-proposal": {
            "get": {
                "description": "GetAllInternalProposal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Proposal"
                ],
                "summary": "GetAllInternalProposal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageIndex",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ItnProposalTitle",
                        "name": "ItnProposalTitle",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateInternalProposal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Proposal"
                ],
                "summary": "CreateInternalProposal",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateInternalProposalDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            },
            "patch": {
                "description": "UpdateInternalProposal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Proposal"
                ],
                "summary": "UpdateInternalProposal",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateInternalProposalDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/internal-proposal/recycle": {
            "get": {
                "description": "GetAllInternalProposalRecycle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Proposal"
                ],
                "summary": "GetAllInternalProposalRecycle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageIndex",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ItnProposalTitle",
                        "name": "ItnProposalTitle",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/internal-proposal/restore/{id}": {
            "patch": {
                "description": "RestoreInternalProposal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Proposal"
                ],
                "summary": "RestoreInternalProposal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/internal-proposal/update-status": {
            "patch": {
                "description": "UpdateInternalProposalStatus",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Proposal"
                ],
                "summary": "UpdateInternalProposalStatus",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateInternalProposalStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/internal-proposal/{id}": {
            "get": {
                "description": "FindInternalProposal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Proposal"
                ],
                "summary": "FindInternalProposal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteInternalProposal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Proposal"
                ],
                "summary": "DeleteInternalProposal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/system-parameter/save": {
            "post": {
                "description": "SaveSystemParameterDto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System Parameter"
                ],
                "summary": "SaveSystemParameterDto",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SaveSystemParameterDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/system-parameter/{sysParaID}": {
            "get": {
                "description": "GetSystemParameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System Parameter"
                ],
                "summary": "GetSystemParameter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sysParaID",
                        "name": "sysParaID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateInternalProposalDto": {
            "type": "object",
            "required": [
                "itn_proposal_content",
                "itn_proposal_title",
                "itn_proposal_type"
            ],
            "properties": {
                "itn_proposal_content": {
                    "type": "string"
                },
                "itn_proposal_title": {
                    "type": "string"
                },
                "itn_proposal_type": {
                    "type": "string"
                }
            }
        },
...
        "model.SaveSystemParameterDto": {
            "type": "object",
            "required": [
                "sys_para_description",
                "sys_para_id",
                "sys_para_value"
            ],
            "properties": {
                "sys_para_description": {
                    "type": "string"
                },
                "sys_para_id": {
                    "type": "string"
                },
                "sys_para_value": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateInternalProposalDto": {
            "type": "object",
            "required": [
                "itn_proposal_content",
                "itn_proposal_id",
                "itn_proposal_title",
                "itn_proposal_type"
            ],
            "properties": {
                "itn_proposal_content": {
                    "type": "string"
                },
                "itn_proposal_id": {
                    "type": "string"
                },
                "itn_proposal_title": {
                    "type": "string"
                },
                "itn_proposal_type": {
                    "type": "string"
                }
            }
        },
        "model.UpdateInternalProposalStatusDto": {
            "type": "object",
            "required": [
                "itn_proposal_id",
                "itn_proposal_status"
            ],
            "properties": {
                "itn_proposal_id": {
                    "type": "string"
                },
                "itn_proposal_status": {
                    "type": "string"
                }
            }
        },
...
        "response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "status code",
                    "type": "integer"
                },
                "data": {
                    "description": "dữ liệu trả về"
                },
                "error": {
                    "description": "dữ liệu lỗi"
                },
                "message": {
                    "description": "thông báo lỗi"
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessTokenAuth": {
            "type": "apiKey",
            "name": "x-at-tk",
            "in": "header"
        },
        "ClientIDAuth": {
            "type": "apiKey",
            "name": "id_user_guest",
            "in": "header"
        },
        "RefreshTokenAuth": {
            "type": "apiKey",
            "name": "x-rf-tk",
            "in": "header"
        }
    },
    "security": [
        {
            "ClientIDAuth": []
        },
        {
            "AccessTokenAuth": []
        },
        {
            "RefreshTokenAuth": []
        }
    ]
}